[REFACTOR] 관리자 권한 & 어르신 정보 접근 권한 확인 매서드 분리
